+++ packages/frontend/src/pages/SqlRunner.tsx
@@ -1,43 +1,28 @@
-import React, {
-    FC,
-    useMemo,
-    useState,
-    useCallback,
-    Dispatch,
-    SetStateAction,
-    useEffect,
-} from 'react';
+import React, { useMemo, useState, useCallback } from 'react';
 import styled from 'styled-components';
-import { Tooltip2 } from '@blueprintjs/popover2';
-import { CopyToClipboard } from 'react-copy-to-clipboard';
 import {
-    Button,
     Card,
     H3,
-    Menu,
-    MenuDivider,
-    MenuItem,
-    NonIdealState,
-    PopoverPosition,
     Text,
     Divider,
     useHotkeys,
+    Callout,
 } from '@blueprintjs/core';
-import { Ace } from 'ace-builds';
-import langTools from 'ace-builds/src-noconflict/ext-language_tools';
 import { TreeNodeInfo } from '@blueprintjs/core/src/components/tree/treeNode';
-import { friendlyName, TableBase } from 'common';
-import AceEditor from 'react-ace';
+import { TableBase } from 'common';
 import { CollapsableCard } from '../components/common/CollapsableCard';
 import { useSqlQueryMutation } from '../hooks/useSqlQuery';
 import { Section } from '../providers/TrackingProvider';
 import { SectionName } from '../types/Events';
 import { RefreshServerButton } from '../components/RefreshServerButton';
-import { BigButton } from '../components/common/BigButton';
-import { ResultsTable as Table } from '../components/ResultsTable/ResultsTable';
 import AboutFooter from '../components/AboutFooter';
 import { useProjectCatalog } from '../hooks/useProjectCatalog';
 import { Tree } from '../components/common/Tree';
+import { useProjectCatalogTree } from '../hooks/useProjectCatalogTree';
+import SqlRunnerResultsTable from '../components/SqlRunner/SqlRunnerResultsTable';
+import RunSqlQueryButton from '../components/SqlRunner/RunSqlQueryButton';
+import SideBarLoadingState from '../components/common/SideBarLoadingState';
+import SqlRunnerInput from '../components/SqlRunner/SqlRunnerInput';
 
 const Wrapper = styled('div')`
     display: flex;
@@ -69,92 +54,16 @@ const CardDivider = styled('div')`
     padding-top: 10px;
 `;
 
-const SideBarLoadingState = () => (
-    <Menu large style={{ flex: 1 }}>
-        {[0, 1, 2, 3, 4].map((idx) => (
-            <React.Fragment key={idx}>
-                <MenuItem className="bp3-skeleton" text="Hello" />
-                <MenuDivider />
-            </React.Fragment>
-        ))}
-    </Menu>
-);
-
-const RunQueryButton: FC<{ isLoading: boolean; onSubmit: () => void }> = ({
-    onSubmit,
-    isLoading,
-}) => (
-    <BigButton
-        intent="primary"
-        style={{ width: 150, marginRight: '10px' }}
-        onClick={onSubmit}
-        loading={isLoading}
-    >
-        Run query
-    </BigButton>
-);
-
-const ResultsIdleState: FC<React.ComponentProps<typeof RunQueryButton>> = (
-    props,
-) => (
-    <Section name={SectionName.EMPTY_RESULTS_TABLE}>
-        <div style={{ padding: '50px 0' }}>
-            <NonIdealState
-                description="Click run query to see your results"
-                action={<RunQueryButton {...props} />}
-            />
-        </div>
-    </Section>
-);
-
-const createCompleter: (fields: Ace.Completion[]) => Ace.Completer = (
-    fields,
-) => ({
-    getCompletions: (editor, session, pos, prefix, callback) => {
-        callback(null, fields);
-    },
-});
-
-export const useProjectCatalogAceEditorCompleter = (
-    sqlTables: string[],
-): {
-    setAceEditor: Dispatch<SetStateAction<Ace.Editor | undefined>>;
-} => {
-    const [aceEditor, setAceEditor] = useState<Ace.Editor>();
-
-    useEffect(() => {
-        if (aceEditor) {
-            const fields = sqlTables.map<Ace.Completion>((sqlTable) => {
-                const technicalOption: Ace.Completion = {
-                    caption: sqlTable,
-                    value: sqlTable,
-                    meta: 'Table',
-                    score: Number.MAX_VALUE,
-                };
-                return technicalOption;
-            });
-            langTools.setCompleters([createCompleter(fields)]);
-        }
-        return () => {
-            langTools.setCompleters([]);
-        };
-    }, [aceEditor, sqlTables]);
-
-    return {
-        setAceEditor,
-    };
-};
-
 const generateBasicSqlQuery = (table: string) =>
-    `SELECT * FROM ${table} LIMIT 25`;
+    `SELECT *
+     FROM ${table} LIMIT 25`;
 
 const SqlRunnerPage = () => {
-    const [copied, setCopied] = useState<boolean>(false);
     const [sql, setSql] = useState<string>('');
-    const [columnsOrder, setColumnsOrder] = useState<string[]>([]);
     const { isLoading: isCatalogLoading, data: catalogData } =
         useProjectCatalog();
-    const { isIdle, isLoading, data, mutate } = useSqlQueryMutation();
+    const sqlQueryMutation = useSqlQueryMutation();
+    const { isLoading, mutate } = sqlQueryMutation;
     const onSubmit = useCallback(() => {
         if (sql) {
             mutate(sql);
@@ -180,69 +89,7 @@ const SqlRunnerPage = () => {
         ];
     }, [onSubmit]);
     useHotkeys(hotkeys);
-
-    const dataColumns = useMemo(() => {
-        if (data && data.rows.length > 0) {
-            return Object.keys(data.rows[0]).map((key) => ({
-                Header: <span>{friendlyName(key)}</span>,
-                accessor: key,
-                type: 'dimension',
-            }));
-        }
-        return [];
-    }, [data]);
-
-    const [catalogTree, autoCompleteSql] = useMemo<
-        [TreeNodeInfo[], string[]]
-    >(() => {
-        if (catalogData) {
-            const sqlTables: string[] = [];
-            const tree = Object.entries(catalogData).reduce<TreeNodeInfo[]>(
-                (accDatabases, [database, schemas], databaseIndex) => [
-                    ...accDatabases,
-                    {
-                        id: database,
-                        isExpanded: databaseIndex === 0,
-                        label: friendlyName(database),
-                        icon: 'database',
-                        childNodes: Object.entries(schemas).reduce<
-                            TreeNodeInfo[]
-                        >(
-                            (accSchemas, [schema, tables], schemaIndex) => [
-                                ...accSchemas,
-                                {
-                                    id: schema,
-                                    isExpanded: schemaIndex === 0,
-                                    label: friendlyName(schema),
-                                    icon: 'diagram-tree',
-                                    childNodes: Object.entries(tables).reduce<
-                                        TreeNodeInfo[]
-                                    >((accTables, [table, nodeData]) => {
-                                        sqlTables.push(nodeData.sqlTable);
-                                        return [
-                                            ...accTables,
-                                            {
-                                                id: table,
-                                                label: friendlyName(table),
-                                                icon: 'th',
-                                                nodeData,
-                                            },
-                                        ];
-                                    }, []),
-                                },
-                            ],
-                            [],
-                        ),
-                    },
-                ],
-                [],
-            );
-            return [tree, sqlTables];
-        }
-        return [[], []];
-    }, [catalogData]);
-    const { setAceEditor } =
-        useProjectCatalogAceEditorCompleter(autoCompleteSql);
+    const catalogTree = useProjectCatalogTree(catalogData);
 
     const handleNodeClick = React.useCallback(
         (node: TreeNodeInfo) => {
@@ -270,24 +117,29 @@ const SqlRunnerPage = () => {
                         }}
                     >
                         <div style={{ flex: 1 }}>
-                            <div style={{ height: '100px' }}>
-                                <div
-                                    style={{
-                                        display: 'flex',
-                                        flexDirection: 'row',
-                                        justifyContent: 'space-between',
-                                        alignItems: 'center',
-                                    }}
-                                >
-                                    <H3>Tables</H3>
-                                </div>
-                                <div style={{ padding: '10px' }}>
-                                    <Text>
-                                        Select a table to populate the sql input
-                                    </Text>
-                                </div>
-                                <Divider />
+                            <div
+                                style={{
+                                    display: 'flex',
+                                    flexDirection: 'row',
+                                    justifyContent: 'space-between',
+                                    alignItems: 'center',
+                                }}
+                            >
+                                <H3>Tables</H3>
                             </div>
+                            <div style={{ padding: '10px' }}>
+                                <Callout
+                                    intent="primary"
+                                    style={{ marginBottom: 10 }}
+                                >
+                                    These tables were detected in your dbt
+                                    project.
+                                </Callout>
+                                <Text>
+                                    Select a table to populate the sql input
+                                </Text>
+                            </div>
+                            <Divider />
                             <div style={{ overflowY: 'auto' }}>
                                 {isCatalogLoading ? (
                                     <SideBarLoadingState />
@@ -315,7 +167,7 @@ const SqlRunnerPage = () => {
                             alignItems: 'center',
                         }}
                     >
-                        <RunQueryButton
+                        <RunSqlQueryButton
                             onSubmit={onSubmit}
                             isLoading={isLoading}
                         />
@@ -324,64 +176,18 @@ const SqlRunnerPage = () => {
                 </Section>
                 <CardDivider />
                 <CollapsableCard title="SQL" isOpenByDefault>
-                    <div
-                        style={{
-                            padding: 10,
-                            position: 'relative',
-                        }}
-                    >
-                        <AceEditor
-                            readOnly={isLoading}
-                            value={sql}
-                            height="300px"
-                            width="100%"
-                            editorProps={{ $blockScrolling: true }}
-                            enableBasicAutocompletion
-                            enableLiveAutocompletion
-                            onChange={(value: string) => {
-                                setSql(value);
-                                setCopied(false);
-                            }}
-                            onLoad={setAceEditor}
-                        />
-                        <div
-                            style={{
-                                position: 'absolute',
-                                bottom: 0,
-                                right: 0,
-                            }}
-                        >
-                            <Tooltip2
-                                isOpen={copied}
-                                content="Copied to clipboard!"
-                                intent="success"
-                                position={PopoverPosition.RIGHT}
-                            >
-                                <CopyToClipboard
-                                    text={sql}
-                                    onCopy={() => setCopied(true)}
-                                >
-                                    <Button minimal icon="clipboard" />
-                                </CopyToClipboard>
-                            </Tooltip2>
-                        </div>
-                    </div>
+                    <SqlRunnerInput
+                        sql={sql}
+                        onChange={setSql}
+                        projectCatalog={catalogData}
+                        isDisabled={isLoading}
+                    />
                 </CollapsableCard>
                 <CardDivider />
                 <CollapsableCard title="Results" isOpenByDefault>
-                    <Table
-                        data={data?.rows || []}
-                        dataColumns={dataColumns}
-                        loading={isLoading}
-                        idle={isIdle}
-                        dataColumnOrder={columnsOrder}
-                        onColumnOrderChange={setColumnsOrder}
-                        idleState={
-                            <ResultsIdleState
-                                onSubmit={onSubmit}
-                                isLoading={isLoading}
-                            />
-                        }
+                    <SqlRunnerResultsTable
+                        onSubmit={onSubmit}
+                        sqlQueryMutation={sqlQueryMutation}
                     />
                 </CollapsableCard>
             </ContentSection>
